+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add support for Quantcast v6: the callback registered with the `addEventListener` command should check for both `eventStatus === 'useractioncomplete'` and `eventStatus === 'tcloaded'`.
 
 ## [2.3.1] - 2020-07-07
 ### Fixed